Reorganize Repository Testing Folders
