+++ src/FluentCMS.Repository.LiteDb.Tests/Repositories/LiteDb/LiteDbAuditRepository_Tests.cs
@@ -3,7 +3,7 @@ using FluentCMS.Tests.DummyEntities;
 using Newtonsoft.Json;
 using Shouldly;
 
-namespace FluentCMS.Tests.Repositories;
+namespace FluentCMS.Tests.Repositories.LiteDb;
 public class LiteDbAuditRepository_Tests
 {
     [Fact]
